Support symlinked paths for repository locations
Some people want to be able to host actual repositories in places like
/var/lib/git for ease of integration with git-daemon. Gitolite3 dropped
support for that, so the only way to accomplish it is via a symlink
(which is the recommended way).

Note that while mostly this works just fine, if you're relying on git
alternates, complex hooks, etc, this can break in weird ways if any
paths are changed (but this is unlikely in 99% of cases).

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>